Project_DATS/AI_project/backend/requirements.txt
google-adk>=0.3.0
google-generativeai>=0.8.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-dotenv>=1.0.0
google-cloud-storage>=2.16.0
pydantic>=2.0.0
